+++ src/app/admin/dialogs/change.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit, Input, Inject } from '@angular/core';
 
+import { MAT_DIALOG_DATA } from '@angular/material';
+
 import { KegomateService } from 'src/app/services/kegomate.service';
+
 
 @Component({
   selector: 'change-dialog',
@@ -14,7 +17,7 @@ export class ChangeDialog implements OnInit {
     newCoffee: {}
   };
   
-  constructor(public kegomateService: KegomateService) { }
+  constructor(public kegomateService: KegomateService, @Inject(MAT_DIALOG_DATA) public data: any) { }
 
   ngOnInit() {
     this.getAvailableCoffees();
