+++ src/app/admin/dialogs/change.html
@@ -1,4 +1,4 @@
-<h2 mat-dialog-title>Changing Keg: keg0</h2>
+<h2 mat-dialog-title>Changing Keg: {{data.keg}}</h2>
 <mat-dialog-content>
   <div class="example-form">
     <mat-form-field class="col-md-12">
