--- src/app/admin/kegs/keg/keg.component.html
@@ -1,13 +1,0 @@
-<div class="keg-details">
-  <div class="name">
-    <br>
-    {{kegDetails.activeKeg | titlecase}}
-  </div>
-  <div class="brand">
-    Currently Serving: {{kegDetails.coffee.name}} by {{kegDetails.coffee.brand}}
-    <br>
-    <button mat-raised-button (click)="openChangeDialog(kegDetails.activeKeg)" class="col-md-2">Change Coffee</button>&nbsp;&nbsp;
-    <button mat-raised-button color="primary" (click)="openAdjustDialog(kegDetails.activeKeg)" routerLinkActive="active" class="col-md-2">Adjust Level </button> &nbsp;&nbsp;
-    <button mat-raised-button color="warn" (click)="openNotifyDialog(kegDetails.activeKeg)" routerLinkActive="active" class="col-md-2">Notifications</button>
-  </div>
-</div>
