--- src/app/admin/kegs/keg/keg.component.scss
@@ -1,8 +1,0 @@
-.keg-details {
-    .name {
-      font-size: 1.2em;
-    }
-    .brand {
-      font-size: .8em;
-    }
-  }
