+++ src/app/admin/kegs/kegs.component.scss
@@ -1,0 +1,8 @@
+.keg-details {
+  .name {
+    font-size: 1.2em;
+  }
+  .brand {
+    font-size: .8em;
+  }
+}
