+++ src/app/app.module.ts
@@ -16,8 +16,7 @@ import { AdminComponent } from './admin/admin.component';
 import { LevelsComponent } from './home/levels/levels.component';
 import { ActivityComponent } from './home/activity/activity.component';
 import { ChartsComponent } from './home/charts/charts.component';
-import { KegsComponent } from './admin/kegs/kegs.component';
-import { KegComponent, AdjustDialog, NotifyDialog } from './admin/kegs/keg/keg.component';
+import { KegsComponent, AdjustDialog, NotifyDialog } from './admin/kegs/kegs.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { GaugeComponent } from './home/levels/gauge/gauge.component';
 import { ChangeComponent } from './admin/change/change.component';
@@ -35,7 +34,6 @@ import { ChangeDialog } from './admin/dialogs/change.component';
     ActivityComponent,
     ChartsComponent,
     KegsComponent,
-    KegComponent,
     NavigationComponent,
     GaugeComponent,
     ChangeComponent,
